+++ cs/src/core/VarLen/IHeapConvertible.cs
