+++ cs/src/core/VarLen/MemoryComparer.cs
