+++ cs/src/core/VarLen/SpanByte.cs
