+++ cs/src/core/VarLen/SpanByteAndMemory.cs
