+++ cs/src/core/VarLen/SpanByteComparer.cs
