+++ cs/src/core/VarLen/SpanByteExtensions.cs
