+++ cs/src/core/VarLen/SpanByteFunctions.cs
