+++ cs/src/core/VarLen/SpanByteFunctions_ByteArrayOutput.cs
