+++ cs/src/core/VarLen/UnsafeLogMemoryExtensions.cs
